refactor v1 fixtures
